Eq constraint for Show instance in AlphaColour
Ignore-this: 98e01cc621effd1003984b8d6edd7d6a

Under GHC 7.4 there is no Eq constrain in Num.  Thus an Eq constrain needs to be added
to the Show instance of AlphaColour because it uses equality to determine if the value
is transparent or not.

darcs-hash:20120117153955-130f7-88a0088024f9fb173e649f711060ba25678d0204